ExactBinding when looking for conversion operators; fixes #3676
